+++ packages/task-graph/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @podley/task-graph
 
+## 0.0.36
+
+### Patch Changes
+
+- Update JavaScriptTask to use input
+- Updated dependencies
+  - @podley/job-queue@0.0.36
+  - @podley/storage@0.0.36
+  - @podley/util@0.0.36
+
 ## 0.0.35
 
 ### Patch Changes